Updated to use Clean Architecture project layout.
