+++ api/src/WebAPI/Controllers/WeatherForecastController.cs
@@ -1,0 +1,22 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace api.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class AccountController : ControllerBase
+{
+    private readonly ILogger<AccountController> _logger;
+
+    public AccountController(ILogger<AccountController> logger)
+    {
+        _logger = logger;
+    }
+
+    [HttpGet(Name = "GetWeatherForecast")]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    public IEnumerable<WeatherForecast> Get()
+    {
+      
+    }
+}
